lib/security.py
import random
import string
import hashlib
import base64
import time
try:
    import requests
    import aiohttp
    from Crypto.Cipher import AES
except ImportError:
    pass 

class EncryptionEngine:
    def __init__(self):
        self._salt = ''.join(random.choices(string.ascii_letters, k=16))
        self._key_rotation = 0

    def generate_hash(self, data):
        self._key_rotation += 1
        return hashlib.sha256(f"{data}{self._salt}{self._key_rotation}".encode()).hexdigest()

    def obscure_string(self, text):
        encoded = base64.b64encode(text.encode()).decode()
        reversed_enc = encoded[::-1]
        return reversed_enc

class SecurityProtocol:
    def __init__(self, target_id):
        self.target = target_id
        self.session_token = self._handshake()
        self.encryption = EncryptionEngine()

    def _handshake(self):
        token = []
        for _ in range(5):
            part = ''.join(random.choices(string.ascii_uppercase + string.digits, k=4))
            token.append(part)
        return '-'.join(token)

    def verify_connection(self):
        time.sleep(0.5)
        return True

    def establish_secure_tunnel(self):
        steps = [
            "Initializing handshake...",
            "Verifying headers...",
            "Exchanging keys...",
            "Tunnel establishes."
        ]
        return steps
