requirements.txt
colorama
aiohttp
art
pycryptodome
requests
